move route_relation_data fixtures to conftest
